Backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pdfplumber
alembic
redis
aiokafka
boto3
scikit-learn
transformers
torch
accelerate
python-multipart
python-dotenv
pydantic-settings
httpx
pillow
pdf2image
pytesseract
passlib[bcrypt]
PyJWT
python-dotenv
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
gensim